moved callAction to reusable util
